+++ srcs/vectors/new_vector.c
@@ -6,13 +6,13 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/15 09:10:18 by cmariot           #+#    #+#             */
-/*   Updated: 2022/05/11 19:16:23 by cmariot          ###   ########.fr       */
+/*   Updated: 2022/05/17 15:12:58 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-t_v3	new_vector(float x, float y, float z)
+t_v3	new_vector(double x, double y, double z)
 {
 	t_v3	new;
